[refactor] Split Profile view into smaller components
