+++ miniPostApp/miniPost/Clean/Presentation/Profile/ProfileView.swift
@@ -16,7 +16,7 @@ struct ProfileView: View {
             VStack {
                 if case let .loaded(userDisplay) = viewModel.state {
                     VStack {
-                        cardView(with: userDisplay)
+                        ProfileCardView(userDisplay: userDisplay)
                         Spacer()
                     }
                     .navigationTitle(userDisplay.name)
@@ -29,40 +29,6 @@ struct ProfileView: View {
         .loadingState(viewModel.state)
         .task {
             await viewModel.fetchInfo()
-        }
-    }
-    
-    private func cardView(with userDisplay: ProfileViewModel.UserDisplay) -> some View {
-        HStack(spacing: 25) {
-            VStack(alignment: .leading) {
-                ProfileInfoRow(iconName: "phone", text: userDisplay.phone)
-                ProfileInfoRow(iconName: "globe", text: userDisplay.website)
-                ProfileInfoRow(iconName: "person.text.rectangle", text: userDisplay.userHash)
-            }
-            
-            Spacer()
-            
-            Image(systemName: "person.crop.circle")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 90)
-                .foregroundStyle(.gray.gradient)
-        }
-    }
-}
-
-struct ProfileInfoRow: View {
-    let iconName: String
-    let text: String
-    
-    var body: some View {
-        HStack {
-            Image(systemName: iconName)
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 25)
-            
-            Text(text)
         }
     }
 }
